Input field and do-feedback components behave as expected now.
The `do-feedback` component used to set its default classes in the
`didReceiveAttrs` hook which was wrong, resulting in double renders
when using it with the `input-field` component.

This could’ve been fixable by moving it to the `init` hook but then the
input field wouldn’t have received the default classes. Unfortunately
the tests were passing either way, because the `config` service was
injected in the input-field-test (but not into the component itself).
This made debugging the issue *not fun*.

All is good now. Moving along.